Add an NPM package for Next.js
