+++ packages/common/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@django-bridge/common",
+  "repository": "https://github.com/django-bridge/django-bridge",
+  "license": "BSD-3-Clause",
+  "main": "src/index.ts"
+}
