+++ packages/common/src/frame.ts
@@ -1,0 +1,11 @@
+import { Metadata } from "./metadata";
+
+export interface Frame<Props = Record<string, unknown>> {
+  id: number;
+  path: string;
+  metadata: Metadata;
+  view: string;
+  props: Props;
+  context: Record<string, unknown>;
+  shouldReloadCallback?: (newPath: string, newProps: Props) => boolean;
+}
