+++ packages/next/src/components/RenderFrame.tsx
@@ -1,7 +1,6 @@
 import React, { ReactElement } from "react";
-import { Frame } from "@django-bridge/common";
 
-import Config from "../config";
+import { Config, Frame } from "@django-bridge/common";
 
 export interface RenderFrameProps {
   config: Config;
@@ -17,11 +16,6 @@ function RenderFrame({ config, frame }: RenderFrameProps): ReactElement {
 
   // Render the view and wrap it with each configured global context provider
   let view = <View {...frame.props} />;
-  config.contextProviders.forEach((provider, name) => {
-    view = (
-      <provider.Provider value={frame.context[name]}>{view}</provider.Provider>
-    );
-  });
 
   // eslint-disable-next-line react/jsx-no-useless-fragment
   return <div key={frame.id}>{view}</div>;
