+++ packages/next/src/index.tsx
@@ -1,0 +1,13 @@
+import {
+  Message,
+  DjangoBridgeResponse,
+  Metadata,
+  Config,
+} from "@django-bridge/common";
+import RenderFrame from "./components/RenderFrame";
+
+export { type Frame } from "@django-bridge/common";
+export type { DjangoBridgeResponse as Response, Metadata };
+export type { Message };
+export { Config };
+export { RenderFrame };
