+++ packages/react/src/index.tsx
@@ -1,10 +1,15 @@
 import React, { ReactElement, ReactNode } from "react";
 
-import Browser from "./components/Browser";
-import { Message, DjangoBridgeResponse, djangoGet } from "./fetch";
-import { Metadata } from "./metadata";
-import { Frame, useNavigationController } from "./navigation";
+import {
+  Message,
+  DjangoBridgeResponse,
+  djangoGet,
+  Metadata,
+  Frame,
+} from "@django-bridge/common";
+import { useNavigationController } from "./navigation";
 import { DirtyFormScope } from "./dirtyform";
+import Browser from "./components/Browser";
 import Link, { BuildLinkElement, buildLinkElement } from "./components/Link";
 import Config from "./config";
 import Form from "./components/Form";
@@ -197,10 +202,8 @@ export type { Navigation } from "./contexts";
 export { DirtyFormContext, DirtyFormMarker } from "./dirtyform";
 export type { DirtyForm } from "./dirtyform";
 export { type NavigationController } from "./navigation";
-export type { Frame } from "./navigation";
-export type { DjangoBridgeResponse as Response, Metadata };
+export type { Frame, Message, DjangoBridgeResponse as Response, Metadata };
 export { Link, BuildLinkElement, buildLinkElement };
-export type { Message };
 export { Config };
 export { Form };
 export { RenderFrame };
